move constants + csv helper to different file
